pyproject.toml
[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["infection"]

[project]
name = "infection"
version = "1.0"
description = "UCC AI Quest 2023"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["pytorch", "template", "deep learning"]
authors = [
  {name = "Minh-Khoi Pham", email = "minhkhoi.pham@adaptcentre.ie" },
  {name = "Nam H Trinh"},
  {name = "Van-Tu Ninh"},
  {name = "Khiem-Tu Le"},
]
maintainers = [
  {name = "Minh-Khoi Pham", email = "minhkhoi.pham@adaptcentre.ie" }
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Deep Learning :: Template",
  "License :: MIT License",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "segmentation-models-pytorch>=0.3.3",
  "lightning>=2.1.0",
  "opencv-python>=4.8.1.78",
  "opencv-python-headless>=4.8.1.78",
  "Pillow>=10.0.1",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "albumentations>=1.3.1",
  "deepdiff>=6.7.1",
  "wandb>=0.16.1",
  "plotly>=5.18.0",
  "hydra-core>=1.3.2",
]

[project.optional-dependencies]
dev=[
    "pytest>=7.2.0",
    "pytest-order>=1.0.1",
    "pre-commit>=2.21.0",
]

[project.urls]  # Optional
"Homepage" = "https://github.com/kaylode/ucc-ai-quest-2023"

[tool.isort]
profile = "black"